Continued to work on 'find_job' which now works as expected. All seems to work now.
Moved example data to the end of the sql script and commented all test inserts/updates in the middle of the file.
Except for the pipeline_processes and pipelines, these are added on table creation.

run.ps1 now starts pipe 1, pipe 2 and pipe 4.
In the example, the pipes should be analysed in the following order: 4, 1, 3. But if a process is finished, it will pick up the lower prio job.

In the python script, I fixed the mark_pipe_job_finished condition. Now every last process of a pipe will mark it complete.