Created a full scale docker-compose.yml.
Apparently it works, I experienced a lot of problems (took me days) with the fluentd process starting up to slowly.
A mandatory property is the `fluentd-async-connect: "true"` option. Without it, it will fail to find the logger and crash the docker deamon when trying to abort the running app. It will then hang on the process that failed to connect with the logging driver.

Committing because it seems stable now.
I will now reset the fluentd.conf to write it's logs to elasticsearch.