Fully working example with 2 docker-compose files. Rewrote example to make use of the `job_queue_interaction/
Due to a race condition with the fluend logger and the docker deamon log driver, I could not get it to work in a stable way within one docker file.

To run:
  docker-compose -f docker-compose1of2.yml up
Once running completely (stdout stopped printing):
  docker-compose -f docker-compose2of2.yml up

When code is changed, run the last statement prepended with a '--build' instruction to rebuild the images, like:
  docker-compose -f docker-compose2of2.yml up --build

You can ignore the messages about orphan processes, docker thinks that the first docker compose containers are orphans because they are missing in the second docker compose file that you are running. Docker checks this by searching for docker compose operations that were ran previously in the same folder. This is based on the following:
```
  -p, --project-name NAME     Specify an alternate project name
                              (default: directory name)
```