Unfortunately, I didn't edit these things recently. Therefore I will do my best to describe what is added.
Added 'Startup_Timestamp', the timestamp of the first heartbeat, to the job_queue_database_schema.sql.
Added elasticsearch interaction to the README.md

I can't precisely explain what was changed to docker-compose1of2.yml!! Please consider this in in production.
the same is the case for fluent.conf.

added pause to job_finder.py so that it will not error-loop